item_service/requirements.txt
Flask
google-cloud-datastore
flask-restful
flask-cors
barely_json
google-cloud-logging
six==1.10